Unify solver and reorganize neural components
