+++ arc_solver/neural/sketches.py
@@ -9,12 +9,12 @@ be used to accelerate search on new tasks.
 from __future__ import annotations
 
 import numpy as np
-from typing import List, Tuple, Dict, Any, Optional
+from typing import Any, Dict, List, Optional, Tuple
 from collections import Counter, defaultdict
 import json
 
-from .grid import Array
-from .dsl import apply_program
+from ..grid import Array
+from ..dsl import apply_program
 
 
 class ProgramSketch:
