+++ tools/train_guidance.py
@@ -18,7 +18,7 @@ sys.path.append(str(Path(__file__).parent.parent))
 
 from arc_solver.grid import to_array
 from arc_solver.features import extract_task_features
-from arc_solver.guidance import SimpleClassifier
+from arc_solver.neural.guidance import SimpleClassifier
 
 
 def load_training_data(challenges_path: str, solutions_path: str = None) -> List[Dict[str, Any]]:
