+++ tools/verify_layout.py
@@ -5,8 +5,9 @@ REQUIRED = [
   "arc_solver/__init__.py",
   "arc_solver/dsl.py", "arc_solver/grid.py", "arc_solver/heuristics.py", "arc_solver/objects.py",
   "arc_solver/search.py", "arc_solver/solver.py", "arc_solver/io_utils.py",
-  "arc_solver/features.py", "arc_solver/guidance.py", "arc_solver/sketches.py",
-  "arc_solver/memory.py", "arc_solver/ttt.py",
+  "arc_solver/features.py", "arc_solver/neural/guidance.py",
+  "arc_solver/neural/sketches.py", "arc_solver/neural/episodic.py",
+  "arc_solver/ttt.py",
   "tools/train_guidance.py", "tools/mine_sketches.py", "tools/build_memory.py", "tools/benchmark.py",
   "models", "data", "notebooks", "scripts", "submission",
   "arc_submit.py", "requirements.txt",
